requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
passlib==1.7.4
python-dateutil==2.9.0.post0
six==1.17.0
Werkzeug==3.1.3
